created functions as utilities for proper refactoring
